Changed: Restructure fs module with feature-flagged blocking/tokio implementations
Split fs.rs into modular implementation with separate blocking and tokio
variants selected via compile-time feature flags.

Changes:
- Split fs.rs into fs/mod.rs, fs/blocking_impl.rs, fs/tokio_impl.rs
- Renamed async_impl.rs to tokio_impl.rs for consistency
- Added compile_error guards for mutually exclusive tokio/blocking features
- Updated all tool implementations to use new fs module APIs
- Added file.flush() in tests to ensure data persistence
- Fixed potential integer truncation in webfetch using usize::try_from
- Cleaned up Cargo.toml dependencies

Benefits:
- Clear separation of sync/async filesystem implementations
- Compile-time feature selection prevents runtime ambiguity
- Safer integer conversions prevent truncation bugs
- More consistent naming across async implementations
