+++ src/llm-coding-tools-serdesai/Cargo.toml
@@ -32,8 +32,6 @@ serdes-ai-models = { version = "0.1", features = [
   "openrouter",
   "huggingface",
 ] }
-serdes-ai-streaming = "0.1"
-futures = "0.3"
 
 # Tool trait is async - async-trait is NOT re-exported from serdes-ai
 async-trait = "0.1"
@@ -56,3 +54,4 @@ tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
 tempfile = "3"
 wiremock = "0.6"
 indoc = "2"
+futures = "0.3"
