+++ src/llm-coding-tools-serdesai/src/task/mod.rs
@@ -7,8 +7,8 @@ use crate::registry::{AgentRegistry, RegistryAgent};
 use async_trait::async_trait;
 use llm_coding_tools_agents::{AgentMode, PermissionAction, Ruleset};
 use llm_coding_tools_core::context::ToolContext;
-use llm_coding_tools_core::tools::TaskInput;
 use llm_coding_tools_core::tool_names;
+use llm_coding_tools_core::tools::TaskInput;
 use serde::Deserialize;
 use serdes_ai::tools::{RunContext, SchemaBuilder, Tool, ToolDefinition, ToolError, ToolResult};
 use std::borrow::Cow;
